BTEX-1098 : For musl-libc integration, the printf is updated for timeval.tv_usec
Reason for change: The tv_usec is `long int` in all the systems but in latest musl-libc it is defined as `long long int`
This leads to a warning on PODs build. Since we treat warning as error, Builds fail.
So, we ensure that the usec is printed as `long long int` for all the platforms to be on safer side.
Also, the rtLog is used only in PODs. All the other RDK devices uses RDKLogger which will never hit this line.
Test Procedure: Build Success
Risks: Low
Signed-off-by: Karunakaran A <Karunakaran_Amirthalingam@cable.comcast.com>
Change-Id: Id8c463af75fe5ece9cd5e534435ab7900687bc69